+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [v0.0.5] - 2026-05-24
+
+### Added
+- Experimental noise finder for Dixon
+- Edge finder function local_std_map function
+- fInd_body_oval function to find body in MRI
+- FOlder comparison by hashing
+
+### Changed
+
+- DaFne segmentation evaluation and comparison notebooks all in Dafne folder, same for with medSAM on top
+- pyproject toml has scipy
+- `.gitignore` extended to exclude more files
+
+---
 ## [v0.0.4] - 2026-05-18
 
 ### Added